Create Company.cs
A CEO of a small company has asked you to create an application that allows him to input employees and the following data about them:
Name, salary, job, department, e-mail address (optional), and age (optional). 
The application must show him which department has the highest average salary and what it is and list the employees working in it after sorting them based on their salary. If the e-mail is missing, it must print "N/A". If the age is missing, it must not print anything. Round up the average salary to two digits after the floating point.