Create Cars.cs
A car-selling company needs software that collects data about their cars (brand, model, engine volume, and year of production) and calculates the annual tax of each car and the total tax of all cars. The annual bill is equal to 20% of the car's engine volume plus: 50$ for cars created after 2010, 60$ for cars produced between 2001 and 2010, and 70$ for cars produced before 2001. The program must add cars until it receives an "End" command that makes it stop and then print all cars sorted by their brand.