Create VechicleDrive.cs
Write a program that stores data for a certain amount of vehicles (model, fuel, fuel consumption, travelled distance). Each car should be a different model (models cannot repeat) and start with a 0 km distance. After you input the data for each vehicle, until the program receives an "end" command, it should obey commands to drive the vehicle to a certain distance and update the vehicle's data or return an error if there is not enough fuel for the drive. After the end command is given, the program will output the updated data for each car.