Create VetCustomers.cs
You are a vet. Write a program that stores the name, age, owner's name, and breed (for dogs only) of the pets you take care of. Your program should save the data about the pets and based on the input it must decide whether the pet is a dog or a cat. Then it should print all pets ordered by their type and age.