Create Family.cs
Create a class called Person, which has fields name and age. And a class, Family, that stores a list of people and has methods to add new members and return the oldest member in the family. Use them to input a family with N members and output the oldest member in it.